refactor: вынести переменные в config
